move helper functions to imageUtils
