+++ app/src/main/java/com/fpf/smartscan/ui/screens/search/SearchResults.kt
@@ -34,51 +34,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import java.io.IOException
-
-/**
- * A simple LRU Cache to hold Bitmaps to avoid decoding them multiple times.
- */
-object BitmapCache {
-    private val cache: LruCache<Uri, Bitmap> = object : LruCache<Uri, Bitmap>(calculateMemoryCacheSize()) {
-        override fun sizeOf(key: Uri, value: Bitmap): Int {
-            return value.byteCount / 1024 // size in kilobytes
-        }
-    }
-
-    private fun calculateMemoryCacheSize(): Int {
-        val maxMemory = (Runtime.getRuntime().maxMemory() / 1024).toInt()
-        return maxMemory / 8 // Use 1/8th of the available memory for cache
-    }
-
-    fun get(uri: Uri): Bitmap? = cache.get(uri)
-    fun put(uri: Uri, bitmap: Bitmap): Bitmap? = cache.put(uri, bitmap)
-}
-
-suspend fun loadBitmapFromUri(
-    context: Context,
-    uri: Uri,
-    targetWidth: Int? = null,
-    targetHeight: Int? = null
-): Bitmap? {
-    return withContext(Dispatchers.IO) {
-        BitmapCache.get(uri) ?: try {
-            val source = ImageDecoder.createSource(context.contentResolver, uri)
-            val bitmap = ImageDecoder.decodeBitmap(source) { decoder, info, _ ->
-                if (targetWidth != null && targetHeight != null) {
-                    decoder.setTargetSize(targetWidth, targetHeight)
-                }
-            }
-            BitmapCache.put(uri, bitmap)
-            bitmap
-        } catch (e: IOException) {
-            e.printStackTrace()
-            null
-        }
-    }
-}
+import com.fpf.smartscan.lib.loadBitmapFromUri
 
 @Composable
 fun MediaStoreImage(
